Improve final function arguments following other complex arguments
Be more conservative about when to format final function arguments
specially. Now only do so if the other arguments are "simple". This
avoids some cases where the current formatting is quite bad indeed. It
might be better to allow the other arguments to be complex as long as
they fit on a line, but that is more involved to implement.