Improve sequences under ifthenelse with unnecessary but safer parens
When the arm of an if-then-else expression ends in a sequence, but
preceded by a let-binding, then parens are not strictly necessary. But
it is an easy mistake to make to edit out the let-binding without
adding parens around the sequence. This diff adds unnecessary parens
in several such situations.

Some examples:
```
;; if true then ( match true with true -> if false then () ) else ()

;; if true then match true with true -> if false then ()

;; if true then (
     let () = () in
     () ; () )

;; if true then ( () ; () )

;; if true then ()
   else
     let () = () in
     () ; ()

;; if true then () else ( () ; () )

;; if true then (
     let () = () in
     () ; () )
   else ()

;; if true then ( () ; () ) else ()

;; if true then match () with () when false -> () ; () | _ -> () ; ()

;; if true then ( if false then () ) else ()

;; if true then
     if false then (
       () ;
       let () = () in
       () )
```