Build: use dune file generation
`dune` cannot directly generate a list of rules depending on things like
file system layout, but there are several workarounds for this.

The first one is "ocaml syntax" - a dune file can be written as an OCaml
script which builds a large string depending on its execution.

The second one is dune file generation: an executable generates a
dune.inc (which is stored in the git repository) which is included in
the main dune file. An extra diff action makes sure that dune.inc stays
up to date.

This switches from ocaml syntax to dune file generation, for the following
reasons:

- ocaml syntax could go away in a future dune version, and does not work
  well with incremental builds
- it makes it possible to use external libraries in the code generation,
  such as stdio in this example.
- it makes debugging easier since it's only a (generated) file to inspect

Some things to note:

- since the file is generated, it needs to be kept up to date. Concretely, it
  means that when a test file is created or delete, it will be necessary to run
  dune promote to update dune.inc. This seems to be rare enough to be workable.
- the file is generated so it's visible in git and makes a large positive diffstat in this cmmit.
- `max-iters` is also set in .ocamlformat, which changes a test that was setting
  `--config` which overrides `OCAMLFORMAT`.