Future-proof Fmt API in case Fmt.t goes abstract
This is a set of three independent improvements to Fmt in case Fmt.t
becomes abstract in the future, but I think they improve the situation
even if that's not the case:

- transform list_pn arguments to required options (from optional ones).
These arguments are always passed, so optional arguments do not have any
advantages. Actually, when Fmt.t is made abstract, this confuses the
argument erasure check since the last argument becomes ?next, which
can't be erased (before, the last argument was the formatter).

- monomorphize ( $ ) and move it to Fmt. It was only used like this in the code base.

- refactor how the output formatter is set up so that it does not use
explicit sequencing. This ensures there is no sequencing with formatters
done in Translation_unit itself. This iteration is done with ( $ )
instead, making sure it will continue working if Fmt.t is abstract.