Remove global ref Cmts.remove (#1142)
This replaces a global reference by a new field in `Cmts.t`. It controls
whether to remove from `remaining` or not.

This mechanism is only used in `preserve`, which did not have access to
`Cmts.t` (as it was referring to the global). So the interface changes a
bit so that the correct value is passed.