Purify Cmts module
As outlined in #1152, this switches the inner representation of Cmts.t
so that it uses maps instead of hash tables.

More precisely, comments tables are encapsulated as tables and their
updates are done using update_x functions which should make the next
step (turning field updates into functional updates) smoother.