Remove some polymorphic comparisons
These can be tricky to remove later on so it is better to isolate them
so that the polymorphic comparison is defined only in one place.

I tried to avoid exposing `is_*` function in the spirit of
[tell-dont-ask](https://martinfowler.com/bliki/TellDontAsk.html) and
preferring pattern matching like in `fmt_direction_flag`, but in some
cases it was not possible.