Only skip tests that require 4.08.0
Some tests can only run on ocaml >= 4.08.0.

Previously, a version constraint on opam would ensure that we do not try
to run any test if the compiler version is not compatible.

It has some drawbacks, in particular ocaml-ci does not have the notion
of a build-only job, so it would skip these versions and we did not have
any build using 4.06 or 4.07.

This implements a different strategy, where we add an `(enabled_if)`
constraint on tests that require a recent compiler. Dune will not attach
the corresponding actions to the global `@runtest` alias, effectively
skipping them.

On the generator side, this works by interpreting files with a
`.minimum-ocaml-version` extension - in this case it creates the proper
`(enabled_if)` option. Note that this means that we use the same dune
files for all compiler versions.

It seems like a lot of files, but in the big picture only few tests
require ocaml >= 4.08.0: 39 out of 217 (or 18%).

Another benefit of this approach is that we run other tests (including
unit tests) on 4.06 and 4.07.