Add doc-comments-val=unset
When `doc-comments-val` was added, the scope of `doc-comments` was reduced and
didn't overlap with the new option.
This is confusing because the name of the option suggest that `doc-comments-val`
overlap with `doc-comments` but not by default.

This also added unecessary maintainance to users of `doc-comments`, they had to
understand and add a new option to their config.

This new value is set by default. This may impact users who kept 0.14.0
defaults when upgrading to 0.14.1 (eg. setting `doc-comments=before` and
not setting `doc-comments-val`).