Fix disabling with attributes on < 4.08 (#1322)
With OCaml < 4.08, attributes do not have a location: they only have a
name (with a loc) and a payload (with a loc). Starting with >= 4.08,
they do, as `attr_loc`.

We use OMP, so we never see the former representation. However, OMP does
not create a new loc or use `Location.none`, it copies the name's loc
into the attribute's loc.

This means that we have two situations:

- >= 4.08: parser provides precise locs:

                        [@@@name "payload"]
    attr.attr_name.loc      ^^^^
    attr.attr_loc       ^^^^^^^^^^^^^^^^^^^

- < 4.08: parser provides loc for attr_name, and OMP copies it:

                        [@@@name "payload"]
    attr.attr_name.loc      ^^^^
    attr.attr_loc           ^^^^

`Source.extend_loc_to_include_attributes`, used to disable formatting on
parts of the AST, tries to determine in which case we are to recover a
better loc in the < 4.08 case.

The existing test would check for `Location.none`, which is never the
case. So the corresponding test was disabled, but with this fix it works
fine.