Build on OCaml 4.11+alpha3 (#1392)
* Do not use Longident.parse anymore

This function from compiler-libs is being deprecated starting with 4.11.
But actually we do not really rely on it, all call sites just parse a
string that can not have any dots. So we expose a smart constructor that
returns a `Lident` from a dotless string.

* Use the latest token type (not compiler-libs ones)

This adds a micro library that copies the token type from OCaml 4.11, so
that when source code is reparsed, the token type from compiler-libs
(which can change) is not directly manipulated.

If the token type from the compiler-libs version is not the latest one, an
upwards migration is performed.