Loc_tree.of_ast: skip locations in doc-comments
The `of_ast` function iterates on all locations and collects them in a
list. There is special code to replace to locations within doc-comments
to `Location.none`. But nothing will be done with these locations, so it
is possible to just cut the recursion by not calling the mapper to
collect locs under there.