Use OMP2 + ppxlib instead of OMP1
Closes #1460

The most visible API change is that `Ast_mapper` (record-based) is
replaced by `Ppxlib.Ast_traverse` (object-based). The port is mostly
mechanical, replacing fields by methods and `default_mapper` by `super`.
Ppxlib has more hooks so it is possible to override `location_stack` in
one go for example.

OMP (like compiler-libs) provides two ways to print ASTs: `Pprintast`
outputs concrete syntax (`let x = 1 in x`) and is provided by ppxlib.
`Printast` outputs a form closer to the `Parsetree` representation
(`Pexp_let (Ppat_var "x", Pexp_const 1, Pexp_var "x"`), but ppxlib does
not provide that. It exposes a s-expression converter, so we use that
instead. This is only used in debug messages.

`Location` is used for two different things: the type and relative
functions are used by ocamlformat as a whole. In ppxlib, the API is
slightly different, so there are minor changes. `Location.input_name`
however is used by the parser, so we need to refer to the one in
compiler libs as `Ocaml_common.Location.input_name`.

`parse-wyc` is a special case because it uses a 4.08 parser, so it
produces a 4.08 AST. The library is split in two between the parser and
associated modules that operate on the 4.08 and the rest that uses the
ppxlib AST. The modified parser inserts generated nodes in places that
would correspond to parse errors. For example, if it determines that
some sequence of tokens can only form an expression (but it's missing
some tokens to do so in a valid way), it will emit `[%merlin.hole]` as
an expression. Later, an AST mapper walks the AST and needs to determine
if an expression is generated or not. This is simple with OMP, but with
ppxlib we use two different AST versions: the parser uses 4.08 (from
OMP) and the mapper uses 4.11 (at the time; more generally, the one
ppxlib uses). So in the Annot module, the mk functions return a 4.08 AST
(they are called by the parser), but the is_generated functions expect a
4.11 one (they are called by a mapper after migration).