Abstract AST iterations in terms of fragment
In several places, we are doing the same kind of processing but on
different bits of the AST. For example normalizing
structures/signatures/use_files, or collecting comments in
structure/signatures/use_file. The current abstraction is to pass
mapping functions to the bits that do the processing, but this has
several drawbacks: this is not very regular (sometimes we need to pass
`mapper -> 'a -> 'a`, sometimes `'a -> 'a -> bool`, etc), and it leaks
mapper in the interface.

This PR adds a type `'a fragment` which represents operations on 'a, and
replaces completely `Translation_unit.t` (which took the dual approach
of putting all operations in a record).

This simplifies many places by only exposing 1 function that operates on
fragments rather than the 3 variants, and in the context of ppxlib
migration this means that we will only have to change the type of
map_ast so that it takes a ppxlib-based mapper instead of an OMP mapper.