Implement the Fmt module on top of a small kernel
Fmt is a lightweight abstraction on top of `Format` boxes (or rather
`Format_` boxes). In its current implementation, we have `type t =
Format.formatter -> unit`, and most of the module relies on that. It
makes the current abstraction pretty leaky (even if it's just within the
module), and hard to experiment with.

This PR defines a small kernel of functions that need to rely on the
formatter abstraction, and implements the rest of the module on top of
that, for example `list_pn`.

One pattern that has emerged is the need for a `lazy_` combinator, that
defers side effects until corresponding bit is actually laid out. A test
is added as an example. This missing combinator the part that was
blocking previous attempts at modifying this module.