Ppxlib: replace some map uses by fold and iter (#1510)
* Ppxlib: replace some map uses by fold and iter

In most places, we are not interested by the return value and just using
`map` for the side effects. These cases can be adapted to use `iter`
instead.

In some places, this removes a good chunk of code that was used only to
build tree fragments that were only ignored later.

Going further, most places only use side effects to build a list or a
map. These can be adapted to use `fold`. This makes state management
more explicit in parse-wyc.