Cache results of Cmts.preserve
The `preserve` function returns a string with the layout of a part of
the AST. It is used in several places to determine if a subpart has a
break for example.

This can be a problem when these constructions are nested, since laying
out the external construction requires laying out the internal ones
first, and this creates quadratic behavior.

This is solved by caching the results of these intermediate layouts.
This cache is unbounded so there are bad corner cases. At worst, each
cached point is associated to the result string which could mean
quadratic memory consumption.

Some ways to fix this:
- use a LRU cache like the one from `core_kernel`
- store only a summary rather than the full string (length, number of
  lines, max indent)

In addition to #1750, a similar case can be triggered using nested
cases. It seems that `expression_width` would be affected but I have not
been able to reproduce that using that code path.

Closes #1750