Allow custom commands for running ocamlformat
Currently it's possible to customize the path or name of the
ocamlformat executable that gets run by the emacs ocamlformat
package. However sometimes it's desirable to run ocamlformat by
running a multi-word command, such as `opam exec ocamlformat --`. This
change adds support running coamlformat with custom multi-word
commands by allowing the `ocamlformat-command` variable to be set to a
list of strings.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>