Sort modes and modalities  (#90)
* Sort modes and modalities

x @ mode1 mode2 and x @ mode2 mode1 are equivalent, as are things like `value
mod mode1 mode2` and `value mod mode2 mode1`. It'll be nice to have a canonical
order on these to avoid thrashing and inconsistency in code with modalities and
mode annotations. I decided to just go with sorting by String.compare for now,
rather than trying to impose any opinions on the order.

Signed-off-by: Aspen Smith <aspsmith@janestreet.com>

* normalize mode order in the extended parser

the current state of the PR leaves the extended parse tree
unnormalized, but this is hidden by the wonderful bug in the
round-trip check

this commit sorts the modes while parsing the extended ast as we are
wont; now we don't need to handle them specially from `Fmt_ast.ml` or
do a separate normalization

Signed-off-by: David Vulakh <dvulakh@janestreet.com>

---------

Signed-off-by: Aspen Smith <aspsmith@janestreet.com>
Signed-off-by: David Vulakh <dvulakh@janestreet.com>
Co-authored-by: David Vulakh <dvulakh@janestreet.com>