fix: handle case where doc does not lex as ocaml (#2684)
* fix: handle case where doc does not lex as ocaml

Fixes #2679

Formatting `mld` files can crash when the input file cannot be lexed as
ocaml. This can happen with LaTeX is found in `{m ...}` or C code in
`{@c[ ... ]}`, for example.

In that case we would crash when building the list of tokens in the
file; however it's not used in the `mld` case.